+++ src/main/java/jdbi_modules/base/StructuredSqlGenerator.java
@@ -34,7 +34,7 @@ public interface StructuredSqlGenerator extends jdbi_modules.SqlGenerator<Struct
      * @return the select part of the query.
      */
     @NotNull
-    default String getSelect() {
+    default String select() {
         return "";
     }
 
@@ -42,7 +42,7 @@ public interface StructuredSqlGenerator extends jdbi_modules.SqlGenerator<Struct
      * @return the part of the cte clause of the query.
      */
     @NotNull
-    default String getCte() {
+    default String cte() {
         return "";
     }
 
@@ -50,7 +50,7 @@ public interface StructuredSqlGenerator extends jdbi_modules.SqlGenerator<Struct
      * @return the part of the from clause of the query.
      */
     @NotNull
-    default String getFrom() {
+    default String from() {
         return "";
     }
 
@@ -58,7 +58,7 @@ public interface StructuredSqlGenerator extends jdbi_modules.SqlGenerator<Struct
      * @return the joins of the query.
      */
     @NotNull
-    default String getJoins() {
+    default String joins() {
         return "";
     }
 
@@ -66,15 +66,15 @@ public interface StructuredSqlGenerator extends jdbi_modules.SqlGenerator<Struct
      * @return the sort orders of the query.
      */
     @NotNull
-    default String getSortOrder() {
+    default String sortOrder() {
         return "";
     }
 
     /**
-     * @return the filters of the query.
+     * @return the filter of the query.
      */
     @NotNull
-    default String getFilter() {
+    default String filter() {
         return "";
     }
 
@@ -85,12 +85,12 @@ public interface StructuredSqlGenerator extends jdbi_modules.SqlGenerator<Struct
                               @NotNull final Set<jdbi_modules.QueryModifier> queryModifiers,
                               @NotNull final Stack<String> modulePrefix) {
         final StructuredSql structuredSql = new StructuredSql(
-                route(modulePrefix, getCte()),
-                route(modulePrefix, getSelect()),
-                route(modulePrefix, getFrom()),
-                route(modulePrefix, getJoins()),
-                route(modulePrefix, getFilter()),
-                route(modulePrefix, getSortOrder()));
+                route(modulePrefix, cte()),
+                route(modulePrefix, select()),
+                route(modulePrefix, from()),
+                route(modulePrefix, joins()),
+                route(modulePrefix, filter()),
+                route(modulePrefix, sortOrder()));
         queryModifiers.stream().map(qm -> structuredSql.applyQueryModifier(qm, queryModifierNameGenerator)).forEach(queryModifierApplier::add);
         return structuredSql;
     }
@@ -102,12 +102,12 @@ public interface StructuredSqlGenerator extends jdbi_modules.SqlGenerator<Struct
                                  @NotNull final Set<QueryModifier> queryModifiers,
                                  @NotNull final StructuredSql sqlType,
                                  @NotNull final Stack<String> modulePrefix) {
-        sqlType.cte += (!sqlType.cte.isEmpty() && !getCte().isEmpty() ? ',' : "") + route(modulePrefix, getCte());
-        sqlType.select += (!sqlType.select.isEmpty() && !getSelect().isEmpty() ? ',' : "") + route(modulePrefix, getSelect());
-        sqlType.from += (!sqlType.from.isEmpty() && !getFrom().isEmpty() ? ' ' : "") + route(modulePrefix, getFrom());
-        sqlType.joins += (!sqlType.joins.isEmpty() && !getJoins().isEmpty() ? ' ' : "") + route(modulePrefix, getJoins());
-        sqlType.filter += (!sqlType.filter.isEmpty() && !getFilter().isEmpty() ? " AND " : "") + route(modulePrefix, getFilter());
-        sqlType.sortOrder += (!sqlType.sortOrder.isEmpty() && !getSortOrder().isEmpty() ? ',' : "") + route(modulePrefix, getSortOrder());
+        sqlType.cte += (!sqlType.cte.isEmpty() && !cte().isEmpty() ? ',' : "") + route(modulePrefix, cte());
+        sqlType.select += (!sqlType.select.isEmpty() && !select().isEmpty() ? ',' : "") + route(modulePrefix, select());
+        sqlType.from += (!sqlType.from.isEmpty() && !from().isEmpty() ? ' ' : "") + route(modulePrefix, from());
+        sqlType.joins += (!sqlType.joins.isEmpty() && !joins().isEmpty() ? ' ' : "") + route(modulePrefix, joins());
+        sqlType.filter += (!sqlType.filter.isEmpty() && !filter().isEmpty() ? " AND " : "") + route(modulePrefix, filter());
+        sqlType.sortOrder += (!sqlType.sortOrder.isEmpty() && !sortOrder().isEmpty() ? ',' : "") + route(modulePrefix, sortOrder());
         queryModifiers.stream().map(qm -> sqlType.applyQueryModifier(qm, queryModifierNameGenerator)).forEach(queryModifierApplier::add);
         return sqlType;
     }