+++ src/test/java/jdbi_modules/base/TestStructuredSqlGenerator.java
@@ -49,31 +49,31 @@ class TestStructuredSqlGenerator {
     private class None implements StructuredSqlGenerator {
         @NotNull
         @Override
-        public String getSelect() {
+        public String select() {
             return "";
         }
 
         @NotNull
         @Override
-        public String getFrom() {
+        public String from() {
             return "";
         }
 
         @NotNull
         @Override
-        public String getJoins() {
+        public String joins() {
             return "";
         }
 
         @NotNull
         @Override
-        public String getSortOrder() {
+        public String sortOrder() {
             return "";
         }
 
         @NotNull
         @Override
-        public String getFilter() {
+        public String filter() {
             return "";
         }
     }
@@ -81,31 +81,31 @@ class TestStructuredSqlGenerator {
     private class Full implements StructuredSqlGenerator {
         @NotNull
         @Override
-        public String getSelect() {
+        public String select() {
             return "A";
         }
 
         @NotNull
         @Override
-        public String getFrom() {
+        public String from() {
             return "A";
         }
 
         @NotNull
         @Override
-        public String getJoins() {
+        public String joins() {
             return "A";
         }
 
         @NotNull
         @Override
-        public String getSortOrder() {
+        public String sortOrder() {
             return "A";
         }
 
         @NotNull
         @Override
-        public String getFilter() {
+        public String filter() {
             return "A";
         }
     }